replaced hamburgermenu with navview in sample app shell
