+++ Microsoft.Toolkit.Uwp.SampleApp/Microsoft.Toolkit.Uwp.SampleApp.csproj
@@ -122,6 +122,9 @@
     <PackageReference Include="Microsoft.Services.Store.Engagement">
       <Version>10.1711.28001</Version>
     </PackageReference>
+    <PackageReference Include="Microsoft.UI.Xaml">
+      <Version>2.0.180916002-prerelease</Version>
+    </PackageReference>
     <PackageReference Include="Monaco.Editor">
       <Version>0.6.0-alpha</Version>
     </PackageReference>
@@ -529,6 +532,8 @@
     <Content Include="SamplePages\InfiniteCanvas\InfiniteCanvas.bind" />
     <Content Include="SamplePages\PlannerTaskList\PlannerTaskListXaml.bind" />
     <Content Include="SamplePages\DataGrid\DataGridCode.bind" />
+    <Compile Include="Shell.Search.cs" />
+    <Compile Include="Shell.SamplePicker.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App.xaml.cs">
@@ -552,7 +557,6 @@
     <Compile Include="Controls\AadAuthControl.xaml.cs">
       <DependentUpon>AadAuthControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Controls\ExtendedHamburgerMenu.cs" />
     <Compile Include="Controls\SampleAppMarkdownRenderer.cs" />
     <Compile Include="Controls\XamlCodeEditor.xaml.cs">
       <DependentUpon>XamlCodeEditor.xaml</DependentUpon>
@@ -937,11 +941,6 @@
     <Page Include="Controls\CodeRenderer.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
-    </Page>
-    <Page Include="Controls\ExtendedHamburgerMenu.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Page>
     <Page Include="Controls\AadAuthControl.xaml">
       <SubType>Designer</SubType>
