+++ Microsoft.Toolkit.Uwp.SampleApp/Pages/SampleController.xaml.cs
@@ -297,7 +297,7 @@ namespace Microsoft.Toolkit.Uwp.SampleApp
                     SidePaneState = _onlyDocumentation ? PaneState.Full : PaneState.Normal;
                 }
 
-                Shell.Current.SetTitles($"{CurrentSample.CategoryName} > {CurrentSample.Name}");
+                Shell.Current.SetAppTitle($"{CurrentSample.CategoryName} > {CurrentSample.Name}");
             }
             else
             {
